more fixes
* move filter_enum to its own file
* random_species -> random_starting species (and ditto for job)
* Don't pick non-starting species/jobs for random/viable selection
