+++ crawl-ref/source/playable.cc
@@ -11,6 +11,7 @@
 #include "json.h"
 #include "json-wrapper.h"
 
+#include "filter-enum.h"
 #include "playable.h"
 #include "jobs.h"
 #include "newgame.h"
@@ -22,19 +23,6 @@
 string combo_type::abbr() const
 {
     return string(get_species_abbrev(species)) + get_job_abbrev(job);
-}
-
-template <typename Enum, typename Predicate>
-vector<Enum> filter_enum(Enum max, Predicate filter)
-{
-    vector<Enum> things;
-    for (int i = 0; i < max; ++i)
-    {
-        const Enum thing(static_cast<Enum>(i));
-        if (filter(thing))
-            things.push_back(thing);
-    }
-    return things;
 }
 
 vector<job_type> playable_jobs()
